+++ src/components/Game.js
@@ -7,23 +7,14 @@ import useInterval from "../hooks/use-interval.hook";
 import cookieSrc from "../cookie.svg";
 import Item from "./Item";
 
-const stuff = [
-  { id: "cursor", name: "Cursor", cost: 10, value: 1 },
-  { id: "grandma", name: "Grandma", cost: 100, value: 10 },
-  { id: "farm", name: "Farm", cost: 1000, value: 80 },
-];
-
-const calculateCookiesPerSecond = (purchasedItems) => {
-  return Object.keys(purchasedItems).reduce((acc, itemId) => {
-    const numOwned = purchasedItems[itemId];
-    const item = stuff.find((item) => item.id === itemId);
-    const value = item.value;
-
-    return acc + value * numOwned;
-  }, 0);
-};
-
-const Game = ({ cookies, setCookies, items, setItems }) => {
+const Game = ({
+  calculateCookiesPerSecond,
+  stuff,
+  cookies,
+  setCookies,
+  items,
+  setItems,
+}) => {
   const incrementCookies = () => {
     setCookies((c) => c + 1);
   };
